Override == and != operators for memory::Allocator class (#1527)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1527

To make use of memory pool for large Allocation we require the Allocator interface which can be used in place of std::Allocator. This new Allocator interface can be used for allocation of vector, strings etc. To enable to do so we need to overload == and != operator as these are required method for allocator interface.

Reviewed By: Yuhta

Differential Revision: D35981516

fbshipit-source-id: 079a6fdc1fa098192452323e0d860a3d749f4a13