Speed up SimpleFunctionAdapter's reuseResult logic (#1609)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1609

Logic was recently added to SimpleFunctionAdapter to reuse one of the input
arguments' vectors as the result vector.  This change makes some small optimizations.

1) Turning a bunch of the conditions into constexprs
Some of the conditions are easily checked at compile time, e.g. is the output type
primitive and fixed width.  In addition, using template functions we can make finding
arguments with types that match the return type into a static check (this was the most
expensive check).

2) We can avoid calling BaseVector::ensureWritable if we're reusing an input. (This is
where we get the most benefit)
This function does 2 things.
a) Reallocates nulls_ if it's not unique, we check isReusableFlatVector which guarantees
nulls_ is unique.
b) Resizes the vector to the maximum of either the current size or the requested size.
We know the input Vectors are at least as big as the result Vector due to the way iterate
over them, so we don't need to do this either.

Reviewed By: mbasmanova

Differential Revision: D36363422

fbshipit-source-id: 68f82702e3cb00dfca85fd41501e2c7224475390