Benchmark Feature Normalization (#1614)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1614

Machine learning is becoming an important use case for Velox, and a lot of work is
going into optimizing Velox for it.

Adding a benchmark for normalizing float features which is a fairly standard operation
and one which we want to make sure we don't regress on.

Results from running in CI

============================================================================
../../velox/benchmarks/basic/FeatureNormalization.cpprelative  time/iter  iters/s
============================================================================
normalize                                                   29.40ns   34.02M
normalizeConstant                                           33.33ns   30.00M

Reviewed By: pedroerp

Differential Revision: D36382333

fbshipit-source-id: d44b7bcf80f1e45f30d31ba4280a71dc21d46466