Introduce AssertQueryBuilder (#1600)
Summary:
There is a proliferation of assertQuery methods in HiveConnectorTestBase and
OperatorTestBase serving different use cases: Plan vs. CursorParameters, splits
or no splits, splits for the only leaf node or for multiple leaves, results are
expected to be ordered or not, results to be compared to DuckDB or to specified
expected values.

We introduce AssertQueryBiulder to allow for specifying different query
execution and verification parameters separately from each other.

A basic usage of the new builder is to run a query and verify results using DuckDB:

```
AssertQueryBuilder(plan, duckDbQueryRunner_)
  .assertResults("SELECT ...");
```

To run multi-threaded, set maxDrivers:

```
AssertQueryBuilder(plan, duckDbQueryRunner_)
  .maxDrivers(5)
  .assertResults("SELECT ...");
```

To compare with expected values:

```
RowVectorPtr expectedValues = ...;

AssertQueryBuilder(plan)
  .assertResults(expectedValues);
```

To run with custom configuration values:

```
AssertQueryBuilder(plan, duckDbQueryRunner_)
  .config(core::QueryConfig::kXxx, "yyy")
  .assertResults("SELECT ...");
```

To run with splits for a single leaf node:

```
AssertQueryBuilder(plan, duckDbQueryRunner_)
  .split(makeHiveSplit(filePath))
  .assertResults("SELECT ...");
```

To run with splits for multiple leaf nodes:

```
AssertQueryBuilder(plan, duckDbQueryRunner_)
  .split(probeScanId, makeHiveSplit(probeFilePath))
  .split(buildScanId, makeHiveSplit(buildFilePath))
  .assertResults("SELECT ...");
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1600

Reviewed By: kgpai

Differential Revision: D36360665

Pulled By: mbasmanova

fbshipit-source-id: ef3f264078871a3a7487db5a142538e1b042a38e