Ensuring evalSimplified() has consistent null semantic (#1616)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1616

Ensure evalSimplified() follows the exact null semantic as evalAll(),
so that they throw exception in the same cases. The difference was
evalSimplified() was only considering and merging null inputs from the top
level expression, and not doing it recursively like evalAll(). This can be
reproduced by the Fuzzer with the command line below:

Reviewed By: spershin

Differential Revision: D36392050

fbshipit-source-id: ec0d75b0f33a8b565519e18d3f05b049255a3667