Change way we calculate output channels for LocalPartition. (#748)
Summary:
X-link: https://github.com/facebookexternal/presto_cpp/pull/748

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1645

Presto Exchange node has 'inputs' field, which dictates the field/column order for each Exchange's source as it is required by the Exchange node itself.
Together with the each source's output layout, it provides the mapping between columns coming from each source to the columns Exchange is expecting.
We need to use that for mapping of the columns from the sources to the columns we output from the Exchange.
Before we used the 1st source output and Exchange's output for this mapping, but the column names aren't guaranteed to match, so mapping fails.

The example query, which fails:
**WITH dw_usage AS (VALUES ('2022-05-09')), mc_usage AS (VALUES ('2022-05-07')) SELECT * FROM mc_usage UNION ALL SELECT * FROM dw_usage;**

In this change we add "inputTypeFromSource" field to the LocalPartitionNode and use that and the source output layout for creating column mapping.

Note that the column mapping would be the same for all sources, so we use only the 1st source for the column mapping.

Reviewed By: mbasmanova

Differential Revision: D36438832

fbshipit-source-id: 68199af46befea6786700bb1c2775376d2ec2606