Add empty java class to fix the test coverages staging deploy
The staging plugin unfortunately doesn't like to be skipped with our
deployment process for release. Instead of adding a skip to deploy or
staging, it would just be better off if it had something empty that it
could deploy. That way if one were to try to run test coverage from
open source, the pom.xml would be there as well.