VectorMaker mapVector taking key & value vectors (#1715)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1715

This diff adds a function to VectorMaker which constructs a MapVector given Vectors of keys and values, along with an std::vector of offsets for where each map begins, and optionally a std::vector of indices for which maps are null.

There is an analogous arrayVector function that takes the elements as a Vector, this heavily borrows from that with the additional condition that the key and value Vectors must be the same size.

Reviewed By: kagamiori

Differential Revision: D36682357

fbshipit-source-id: f1579b4e633388f15170ceb8f0209c603247c605