Refactor SeekableInputStream::loadIndices() (#1723)
Summary:
loadIndices() in SeekableInputStream returns the the positions array
index for the next stream in the current column for DWRF, relative to
the startIndex parameter. It was used in
SelectiveStringDictionaryColumnReader to lazily load the stride
dictionary for DWRF. To prepare for the upcoming refactoring that
moves this class and its subclasses to dwio::common, we need to decouple
the DWRF related concept from it. This commit changes loadIndices() to
positionSize() which returns the number of position values this input
stream uses to identify an ORC/DWRF stream address.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1723

Reviewed By: abhash09

Differential Revision: D36781057

Pulled By: zzhao0

fbshipit-source-id: 054157a6e60649a389217aedd96e218dace1ffc9