Add read hints to ExtendedFileSystem open API
Large files are stored in the blob storage. Each worker only processes
a part of the large file. Before this change the open call, has no hint
on what part of the file worker is going to work on. This change adds
the startOffset and length hint, which can be used by a file system
to open only the part of the file.