Add hdfs read support (#1427)
Summary:
Add Hdfs read support. The design is similar to the s3 connector.
There is an HdfsFileSystem class that creates the client and vends HdfsReadFiles.

HdfsReadFile is in charge of returning data from the file via the pread methods. On each pread HdfsReadFile creates an hdfsFile and reads the requested data, after the read is finished the file is closed. Since each pread creates its own file object, the pread methods are thread safe.

To test this I depend on the hadoop executable that has a local mini cluster implementation. For my tests I start a mini cluster and read from it.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1427

Reviewed By: mbasmanova

Differential Revision: D36914786

Pulled By: pedroerp

fbshipit-source-id: d954a2e5692fe922b350318e8c667f1324ee9b83