Refactor integer readers to share boilerplate (#1770)
Summary:
The template instantiation logic for direct and dictionary integer
readers is the same. The difference is at the top of read() where we
seek different streams and at the bottom where we call different
decoders on different visitors.

Adds a conversion from direct to dictionary visitor. This supports
switching between encodings in mid-stripe.

This allows directly inheriting the integer logic for Parquet and Alpha.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1770

Reviewed By: zzhao0

Differential Revision: D37180553

Pulled By: oerling

fbshipit-source-id: 74a45a6d6ef3e6389bb44fc430bb2bd9b6cc9b83