Clarify BaseVector::copy(...) error message (#1899)
Summary:
BaseVector::copy(...) was a virtual function with a default implementation of VELOX_NYI(). Several of its derived classes had implemented this method, so it could be used successfully in many cases. This was confusing.

This PR clarifies its error message to indicate that the operation is supported for Flat vectors.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1899

Reviewed By: kagamiori

Differential Revision: D37535858

Pulled By: mbasmanova

fbshipit-source-id: d4d61f4daab1fbfa42ca158f16b310cf0ab0b673