Print descriptive error message on benchmark jobs (#1887)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1887

Adding a more descriptive message to users in case main builds are broken. Benchmark runs for PRs compare with a baseline (based on main), so in case main is broken the benchmark regression test suite fails. Printing a more descriptive error message to alert users about what's happening, and asking them to rebase on a healthy main.

Reviewed By: kagamiori, mbasmanova

Differential Revision: D37508791

fbshipit-source-id: b26b8fcbb67e98424045e37442fd878155b69e77