Annotate PoS RDDs with PlanFragment ID
PoS internally converts a Presto physical plan (DAG of PlanFragments)
into a DAG of Spark RDDs which is then executed on Spark Cluster.
There is 1:1 mapping between a PlanFragment and an RDD in the DAG.
Currently, there is no way to map an RDD back to a planFragment
and this makes debugging difficult. RDD provides an API to specify
a name for the RDD which is then shown in Spark UI. We can leverage
this API to associate Presto's plan fragment ID with an RDD in Spark
DAG and this can help in tracking which Spark stage corresponds to
which planFragment in Presto plan.