Query failing with exceed local memory doesn't free memory correctly
Query with HashBuildOperator, if fails with exceed local memory doesn't free memory for the operator.
This leads to general pool keeping the memory for it and makes the amount of memory not usable by other
running queries.
Reason for the leak, we reserve the memory and then do memory constraint check. And if query fails due to memory
constraint, we don't free the reserved memory. We changing the order so first do memory constraint check to address it.