Optimize KLL sketch memory usage for small groups (#1819)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1819

When the sketch is created, currently we allocate the memory depending
on the K passed in.  When the actual number of elements being tracked is much
smaller than K (often happening with group by aggregation), the unused memory is
wasted.  We optimize for this case by allowing the actual retained elements to
grow from empty instead of an initial size of K.

We also replace `mt19937` with `default_random_engine` to save some overhead memory for each object.

Reviewed By: mbasmanova

Differential Revision: D37182402

fbshipit-source-id: 4e413a88a1827cdbae85cda960fa51a91f93184b