Add ApproxMostFrequentStreamSummary to calculate approx_most_frequent (#1741)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1741

Create the data structure and its unit test and benchmarks.  The data structure here is a variance of *Stream-Summary* created by [A. Metwally, D. Agrawal, and A. Abbadi](https://www.cse.ust.hk/~raywong/comp5331/References/EfficientComputationOfFrequentAndTop-kElementsInDataStreams.pdf).  It keeps track of values and their counts, and if the number of tracked values exceeds the capacity, we replace the value with least counts with the new value.  The same implementation is used in Presto Java.

Reviewed By: mbasmanova

Differential Revision: D36778577

fbshipit-source-id: 408c679527768b3a1f83cae6dd0a8990634dfa2c