Implement NegatedBytesValues filter class (#1893)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1893

This diff adds support for strings (`varchar`) as a supported type for `NegatedValues` filters. Previously, these filter classes supported integers only, however, queries of the form `NOT IN [list]` for other types would still use the default `MultiRange` filters. Code for the `NegatedBytesValues` class forms a wrapper around `BytesValues` objects, and is contained in `Filter.h` and `Filter.cpp`, except for the `MergeWith` functions which will be added in a future diff.

To create unit tests, an additional set of tests has been added to `FilterTest.cpp`. An additional support function for these tests has been added to `ExprToSubfieldFilter.h`, which is a version of `notIn` for vectors of string values rather than vectors of integer values.

Reviewed By: gggrace14

Differential Revision: D37325631

fbshipit-source-id: d910a0d8e8a096d2146eebce970e1e95c23db593