Fix rowcount generation for lineitem (#1935)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1935

Lineitem is generated based on rows counts from the Orders table, so
the offsets need to advance based on the requested rows, rather than on number
of returned rows in the vectors. This is more general and works for other
tables as well, since they only case when less rows are returned than requested
is when the connector exhausted the generated rows.

Reviewed By: mbasmanova

Differential Revision: D37584637

fbshipit-source-id: 642f6f553464f0a730444e22aa1396295a6a127f