Add NegatedBytesValues filters to dwrf reader (#1960)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1960

This diff adds support for the `NegatedBytesValues` filter class to the dwrf reader, which takes a filter and reads rows from a file that are accepted by the filter. These changes were made in the two files that deal with reading values of type varchar, which are `SelectiveStringDictionaryColumnReader.h` and `SelectiveStringDirectColumnReader.cpp`.

With this added support, `FilterGenerator.cpp` has also been modified to sometimes create `BytesValues` and `NegatedBytesValues` filters (similar to what was done in D37133348 (https://github.com/facebookincubator/velox/commit/f2ec094803f92badf100b4c20d25cfa6fa8c2e08)) so that the unit tests in `E2EFilterTest.cpp` will check this functionality as well.

Reviewed By: gggrace14

Differential Revision: D37398220

fbshipit-source-id: 2546f8ac5b8f2dcd12229c85ee665368e907ee39