Fix buffer reuse bug on Upper/Lower function (#1961)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1961

Another bug caught by Fuzzer: UpperLower vectorized function was
moving away the input buffer and losing all references to it, causing the
Vector that holds the string buffer to run out of scope and be destroyed before
the function execution started, causing the function to trigger a
use-after-free ASAN crash. The issue is only reported in ASAN mode.

Reviewed By: Yuhta

Differential Revision: D37674304

fbshipit-source-id: 2af4945433d8320b61622e38ca5fc83f34712e8a