SimpleFunction should only reuse variadic arg vector if it's flat and reusable (#1940)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1940

This fixes a bug where we wouldn't check to see if a vector in a Variadic argument was
reusable and flat before reusing it. (Thanks Laith for catching this!)

Added a check so we use the first reusable flat vector in those passed under the
Variadic type.

I also did some refactoring of tests so that we test for reuse in SimpleFunctionTest
instead of ExprStatsTest (this isn't really testing the ExprStats but the behavior of
SimpleFunction Adapter).

Reviewed By: laithsakka

Differential Revision: D37636313

fbshipit-source-id: 80438f469ae99a0d24e25cae1cf3650467f4c9f8