Pivoting task, used std::move() to avoid copies
Summary:
We cannot reuse the result vector without modifying the vector interface (we need something like setKeys and setValues that will modify raw* flavors of member variables too). This would be expose too much of the internals and will require more discussion.

The previous changes have been reverted, but we pivoted the task to improve assignment of result with std::move() to avoid copying the shared pointers.

Differential Revision: D37727762

fbshipit-source-id: 80a69f4f362d4e3446b7a95bbc19e77c5063d184