Add allocation tag to the "delta" in memory limit error message
With this change we would be able to know which operator is requesting more memory in the error message. For example, we know the culprit is HashBuildOperator in the following example error.
```
"Failure": "Query exceeded per-node user memory limit of 5GB [Allocated: 624B, Delta: 11.59GB (HashBuilderOperator), Top Consumers: {HashBuilderOperator=23.14MB, GenericPartitioningSpiller=3.39kB
```