Add TPCH query 19 to TpchQueryBuilder (#1933)
Summary:
Adds TPC-H query 19 to TpchQueryBuilder.
Extends TpchBenchmark and ParquetTpchTest with query 19.
```
| # Num Threads/ Drivers | Velox(seconds) | DuckDB(seconds) |
|:----------------------:|:--------------:|:---------------:|
|            1           | 6.35           | 9.50            |
|            4           | 2.38           | 5.85            |
|            8           | 1.55           | 3.05            |
|           16           | 1.39           | 1.55            |
```

At the top of the Velox profile are `dwrf::IntDecoder::bulkRead` and
`dwrf::IntDecoder::bulkReadRows`. These relate to`varint` decoding
and are specific to the DWRF format.
The third is `exec::ConjunctExpr::UpdateResult` which is expected for this query.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1933

Reviewed By: pedroerp

Differential Revision: D37823070

Pulled By: kevinwilfong

fbshipit-source-id: 954edcb94faf3c52cfbcf2cc313162aab4444cb5