Create NegatedBigintRange filter class (#1966)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1966

This diff adds the `NegatedBigintRange` class of filters, which represents queries in the form `NOT BETWEEN low AND high` for big integer bounds. The necessary methods for a big int filter have been added to `Filter.h` and `Filter.cpp`, which builds a wrapper class around the existing `BigintRange` filter for BETWEEN queries. This diff also adds tests for the new class and its member functions within `FilterTest.cpp`, and modifies `ExprToSubfieldFilter.h` to create methods capable of generating these new filters for testing.

Reviewed By: Yuhta, gggrace14

Differential Revision: D37496231

fbshipit-source-id: f2654533036073072bb799c617c1614e22e7f566