Added handling for failing seed problem in testFlatmapAsMapFieldLifeCycle
Summary: The test was failing because with the random seed = 255185196, we happened to get no columns that had any nulls in the random table generation. There is an optimization that sets the nulls buffer to null if there are no null elements to be specified. In this case, the test was failing, so I adjusted the expect statement to reflect this case.

Reviewed By: abhash09

Differential Revision: D37901524

fbshipit-source-id: ba755f78678817893c777e795eb78ca9fdfedb20