Implement MergeWith for NegatedBigintRange filters (#2038)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2038

This diff implements the `MergeWith` function for the newly added `NegatedBigintRange` filters, which allows two filters to get merged into a larger filter which only accepts values that pass both of the original filters. `MergeWith` for the `NegatedBigintRange` filters is compatible with untyped (null/not null/true/false) filters as well as other filters on integers, such as `BigintRange`, `BigintMultiRange`, and the `Values` and `NegatedValues` classes of filters. The corresponding `MergeWith` methods in the other integer classes have also been updated to support merging with `NegatedBigintRange` as well.

Tests for `MergeWith` on `NegatedBigintRange` filters have been added to `FilterTest.cpp` as well.

Reviewed By: Yuhta, gggrace14

Differential Revision: D37503903

fbshipit-source-id: 32137736918c7de88fbd64199e73dd32d68cd6b9