Drop SelectiveColumnReader inheritance from ColumnReader (#2003)
Summary:
SelectiveColumnReader will be moved to dwio::common but ColumnReader
will remain untouched, based on the assumption ColumnReader and its
descendent classes will retire soon. This commit removes the inheritance
of SelectiveColumnReader to ColumnReader. To do this, DwrfRowReaderShared
would contain a SelectiveColumnReader and a ColumnReader as its members,
but only one of them is not null. TestColumnReader.cpp was also modified
this way to make the tests work. In addition, the contents in
TestColumnReader.cpp were reordered to form 3 sections: 1) test class
definitions, 2) tests, and 3) test instantiations.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2003

Reviewed By: oerling

Differential Revision: D37929018

Pulled By: Yuhta

fbshipit-source-id: 3eb3a8115ea146fa58c4e9ee6ee87080f6eaee91