Integrate NegatedBigintRange filters into CreateBigintValues (#2049)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2049

This diff integrates the new `NegatedBigintRange` filter into the `CreateBigintValues` function, which is responsible for taking a list of values to either accept or reject and create the corresponding filter. Rather than using a `MultiRange` filter to represent the case where all values in a range (and none others) are rejected, we instead use the `NegatedBigintRange` filter. This also applies to a range of size 1, which is the `<>` or `!=` case.

Some of the tests in `FilterTest.cpp` have been correspondingly modified to reflect the change from `MultiRange` to `NegatedBigintRange` filters.

Reviewed By: Yuhta, gggrace14

Differential Revision: D37536173

fbshipit-source-id: 844ac265220675c01ef93bcff9c83de1fa2a3708