Add benchmarks for NegatedBigintRange filter (#2052)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2052

This diff adds benchmarks for the newly added `NegatedBigintRange` filters, which handle queries of the form `<> [value]` and `NOT BETWEEN [value 1] AND [value 2]` for integer types. The new filter is tested against the previous `BigintMultiRange` implementation for two types of tests. The first group runs a `<> 0` filter against a list of values, where the percentage of `0` values varies. The second group runs range filters of varying sizes against a set of random values between `-20000` and `20000`.

The results can be found at P513797040. They indicate that the new class of filters is more efficient than the previous `BigintMultiRange` implementation, with performance gains increasing as rejections occur more frequently. In particular, `<>` queries seem to be the biggest winners going over 2x faster than previous, while the `NOT BETWEEN` queries only run about 1.2x faster. Since the `BigintMultiRange` filter was already constant-time with only two ranges, this is expected.

Reviewed By: gggrace14

Differential Revision: D37606511

fbshipit-source-id: 6e58038bd29cc5c371e914c29f1301e3f19b29c3