Implement MergeWith for NegatedBytesRange filter (#2064)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2064

This diff implements the `MergeWith` method for `NegatedBytesRange` filters. Since merging most negated ranges will result in the creation of a `MultiRange` filter that cannot be easily represented otherwise, many of these merges are performed by creating the equivalent `MultiRange` filter with the `toMultiRange` method. This allows the `AND` merge of the new `NegatedBytesRange` filter with another varchar or untyped filter. Tests for this have also been added to `FilterTest.cpp`.

Reviewed By: Yuhta

Differential Revision: D37762978

fbshipit-source-id: a63386ab46b80d2b683702240fc4a3fc85992500