Add a pool of recyclable vectors (#2069)
Summary:
ML data preprocessing workloads perform simple calculations over small batches
of flat vectors. The overhead of allocating vectors for intermediate results in
these workloads is significant. This change adds a thread-local pool of recyclable
vectors. This pool will be used to reduce allocations during expression evaluation.

VectorPool holds up to 10 recycled vectors for each primitive type.

VectorPool::get(type, size) returns a vector of the specified type and size. If
a recycled vector of the same type is available, it is resized and returned.
Otherwise, a new vector is allocated from the memory pool passed in the
constructor.

VectorPool::release(vector) checks if the vector is recyclable, e.g. flat and
signly-referenced, and std::moves it to the pool. If vector is not recyclable
it is left in place. Release operation is best effort and returns a boolean
indicating whether the vector was successfully moved into the pool.

VectorPool::release(vectors) is a convenience method to return multiple vectors
to the pool at once. Only a subset of recyclable vectors is moved into the
pool. The rest are left in place. This operations returns the number of vectors
successfully moved into the pool, which can be 0.

Complex type vectors do not support recycling. These are always allocated from
the memory pool.

Before this change, ExecCtx hosted pools of SelectivityVectors and
DecodedVectors. Now, ExecCtx also hosts a VectorPool.

VectorPool stored in ExecCtx is also accessible from an EvalCtx.

BaseVector::ensureWritable now takes an optional VectorPool parameter.

A new EvalCtx::ensureWritable method is a convenient shortcut for
BaseVector::ensureWritable.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2069

Reviewed By: xiaoxmeng, oerling

Differential Revision: D38035431

Pulled By: mbasmanova

fbshipit-source-id: 2b46f4fce8b7749366fb53d127f0bda5a68a70e0