Fix null dereference in SwitchExpr that has no else and no match (#2082)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2082

The result vector of a SwitchExpr is firstly create when a branch of the switch gets executed. If no branch matches any row, the result vector remains a nullptr. Meanwhile, if a SwitchExpr has no "else" clause, the rows that do not match any branch should be set to null. Therefore, there is a corner case where a SwitchExpr has no "else" clause and no row match any branch. In this situation, the result vector is nullptr when all rows are set to null, which caused null dereference. This diff fixes this bug by creating the result vector before setting nulls.

Reviewed By: mbasmanova

Differential Revision: D38051577

fbshipit-source-id: d9fb09119c60aa3b3b84afffc01df9406a386359