Compact KllSketch to reduce serialized byte size (#2037)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2037

When the value cardinality is low, there are many duplicate items in
lower level, which we could move to upper level to save some space.  This change
implements that with a compact() method, which can be called before
serialization.

After compact (or for a sketch generated by `fromRepeatedValue`), there
is no room in the bottom level, but all levels can still be under-populated.  If
we want to insert into such a sketch, `findLevelToCompact` will not be able to
find a level to shrink.  This fix enables it to first grow the bottom level in this case.

Reviewed By: tanjialiang

Differential Revision: D37925401

fbshipit-source-id: f475ceb85119ccd1f231fca057029e478726ac0e