DiscoveryNodeManager to not return Coordinator/Resource Manager if it's not in the allowed list
In some mysterious conditions, we noticed coordinators are showing up in discovery service, and this lead to
a bad state where resource managed start failing getResourceGroupInfo call due to incosistency in the coordinator
count. we won't to mitigate it by removing this checks which are no longer needed given NodeStatusService should filter
them.