Fix hashcode and equals on Block
Hashcode and equals on blocks are non deterministic, this caused NPE in Planner.

```
java.lang.NullPointerException
	at com.facebook.presto.sql.planner.optimizations.ApplyConnectorOptimization.optimize(ApplyConnectorOptimization.java:154)
	at com.facebook.presto.sql.planner.LogicalPlanner.plan(LogicalPlanner.java:196)
	at com.facebook.presto.sql.planner.LogicalPlanner.plan(LogicalPlanner.java:184)
```

FilterNode -> RowExpression(ConstantExpression) -> value is a block.
Went through all blocks implementation of the hashcode and equals and fixed it.

Will work on improving the ApplyConnectorOptimization to use identityHashCode as
PlanNodes are immutable objects.