Turn ExprEncodingTest into parameterized test (#2127)
Summary:
Break up ExprEncodingTest::encodings test into smaller test each evaluating a
single expression. Parameterize the test on the encodings of the inputs and a
boolean indicating whether to use lazy vectors or not. Remove encodingsOveLazy
test as it is now covered by the boolean flag.

The new test makes it easy to understand which particular expression is failing
and for which set of encodings. It is now easy to re-run just the failing
expression.

The new test also provides better coverage for lazy vectors and exposed a number
of bugs that need to be fixed in follow-up PRs. For now, the failing tests are
disabled.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2127

Reviewed By: bikramSingh91

Differential Revision: D38218005

Pulled By: mbasmanova

fbshipit-source-id: 1fd79176d1a9c54e811b8c73e96a3598ad9c2c73