Fix RowVector::copy() when source is a null constant (#2170)
Summary:
RowVector::copy() was crashing when copying a null constant vector of size > 64.

When decoding null constant,
- DecodedVector::nulls() returns a pointer to a 64-bit buffer
  (&constantNullMask_) regardless of the size of the vector;
- DecodedVector::nullIndices() returns nullptr which indices that to access null
  flag in 'nulls' buffer one should use the top-level row number directly:
  `nulls[row]`

Since nulls buffer has only 64-its, accessing null bit for row number > 64 is
not valid and either causes a crash (if running with ASAN) or incorrect
results.

A fix implemented here is to
- document that nullIndices() is available only when isConstantMapping() ==
  false and add a check for that;
- update RowVector::copy to use DecodedVector::nullIndex() instead of
  nullIndices.

Updated VectorTest to exercise the code path that used to trigger the bug.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2170

Reviewed By: xiaoxmeng

Differential Revision: D38331254

Pulled By: mbasmanova

fbshipit-source-id: 059f156430fd6146182f79920437475dc7c709cf