Add run name to conbench runs (#2209)
Summary:
In https://github.com/facebookincubator/velox/issues/2193 we added pushing to conbench, but I should have also added the run name and run reason. The name is needed (currently) for calculating history of the main branch. After this is merged we should start seeing the history plots on runs.

I've also added a small fix for a typo in the runner name.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2209

Reviewed By: Yuhta

Differential Revision: D38425865

Pulled By: pedroerp

fbshipit-source-id: e940c46064c93c9bf329d8b3e8bf64b9ee49e1b5