Do not copy StripeMetaDataCache contents (#1927)
Summary:
Do not copy StripeMetaDataCache contents, re-add CacheInputTest.cpp

When data comes from a CacheInputStream, the the DWRF metadata at the
end of the file is normally read into cache. This data was then copied
into a new contiguous block of memory for each stripe of the file. In
one case this would be 6.5MB each time. This would be more than all
other memory allocation combined and the minor faults this caused
would drive the system CPU to be over half of total CPU. Fixing this
drops system CPU down to ~20% of total.

Now, instead of copying the metadata, the StripeMetadataCache gives
out a CacheInputStream that is scoped to a range of bytes inside the
cached data. Because the stream is used for reading protobufs, the
stream must end after the serialized metadata, otherwise the protobuf
decoder will read past end. Hence we add an interface to clone
CacheInputStreams and to limit them to a subrange of the originally
covered range of bytes.

The code is covered vy existing tests in TableScanTest.cpp, which all
access this metadata.

Adds a test for the new functions of CacheInputStream.

Enables the dwrf/test/CacheInputTest.cpp. In so doing, removes extra
logging of eviction in non-checkpointed SSD cache mode, since this
makes noise in the error log.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1927

Reviewed By: Yuhta

Differential Revision: D38283254

Pulled By: oerling

fbshipit-source-id: 274fbbcf27033a49c7898001dac79d995744046c