Add arrow conversion for array and map (#2081)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2081

In order to gap the difference between offset representation between velox and arrow, we introduce a `Selection` filter representation that can keep track of which the rows selected and their order.  This way we can compact and reorder the rows in correct order for the entire subtree, which is required by Arrow format.

Reviewed By: kevinwilfong

Differential Revision: D38036430

fbshipit-source-id: 7a550e3843e7a9a0de36fd91e90856a42332957e