Use right intermediate type to init Accumulator for float sum (#2183)
Summary:
Fix the float sum aggregation by using the right type to initialize the accumulator. The current code uses the result type to initialize the accumulator for float sum which is incorrect. As the result type is float, however the actual accumulator type used to store the intermediate result is double. When we cast a double to float and initialize to zero, we only init the h32 part and l32 part is unchanged, we can easily reproduce this bug in SumDouble by re-running the test iterator multiple times and we see the float sum result just keep bumping up by test iterations: 1 X actual result, 2 X actual result, ...

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2183

Reviewed By: kgpai

Differential Revision: D38359075

Pulled By: xiaoxmeng

fbshipit-source-id: 6010f107a0250168459299509318c7b61d7ef3c0