Relax exception comparison in expression fuzzer and allow disabling constant-folding in common path from CLI (#2220)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2220

For some expressions, the common and simplified evaluation
path may process input rows in different orders because the
common path peels off input encodings while the simplified path
flatten input vectors. As the result, the two paths could encounter
different bad inputs first and hence throw different exceptions.
We have seen this happening for a few times with unmatched
error code of UNSUPPORTED and INVALID_ARGUMENT
between the two evaluation paths. Therefore we intentionally
relax the exception comparison to not distinguish between these
two error codes.

In addition, this diff adds a parameter to the expression fuzzer to
control whether to enable constant-folding in the common path.
It's enabled by default. This is needed for the fuzzer test for an
internal engine.

Reviewed By: pedroerp

Differential Revision: D38465045

fbshipit-source-id: 87c2e61da960b280ecebaec30a692247427f786e