Add Parquet Integer Dictionary (#2201)
Summary:
Adds a Parquet RLE  decoder and reading of dictionary page headers.

Adds a special case of non-null that is recognized as a RLE of all ones in single-bit definition levels.

Fixes cases of reading nulls only, e.g. with multiple is null
filters. The generic FormatData::skipNulls gets a flag to indicate if
this is called for a column where only null flags are of interest or
whether this is a part of reading a column with multiple streams.

Does not initialize

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2201

Reviewed By: Yuhta

Differential Revision: D38399637

Pulled By: oerling

fbshipit-source-id: d4c26ab3b4e94ec4ace5e521485c1f6f5e200a48