Override hashcode and equal functions for block builder
Custom hashcode and equal functions are introduced for xxxBlock in PR #17720, where content in the corresponding
block rather than the object itself are used in hashcode and equal function computation. However, xxxBlockBuilder
which also extends from Block interface is left behind. In this commit, we override such functions for block builder.