Add utility docker and docker-compose way of building VELOX (#2181)
Summary:
This PR tries to add a utility docker and docker-compose way of building velox and running tests.
As a user I don't want to install system dependencies on my local machine. I would like to be able to build VELOX on a Docker container.

This PR requires the changes on:
https://github.com/facebookincubator/velox/pull/2177
and
https://github.com/facebookincubator/velox/pull/2180

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2181

Reviewed By: Yuhta

Differential Revision: D38541909

Pulled By: kgpai

fbshipit-source-id: 67e70e321d2b3d04dbf40b514ade9caca76aa7ad