Reduce memory usage of StreamDescriptor
StreamDescriptor stores stream name, field name, type, datasource list
of nested stream descriptors. For Flat Maps, there is lot of reference
and it consumes lot of memory.

This change applies the database schema normalization and refactors the
code so that only streamId and sequence id are stored in the stream
descriptors. Others are stored in common all streams so that the
data will be shared.

This might slightly regress the memory for normal tables, but will save
a lot for flat map tables.