Getting struct writing tests working and using new config for keys (#2248)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2248

Fixed defaulting of non-empty map {{}} bug in testMapWriter. Also fixed a bug where preceding optional arguments were not set. Added pBatches pointer to iterate either original batches or structs depending on what the input to the writer will be. Added structKeys_ member to hold the vector of keys for the column that the writer is created for. parseKeys() converts from string to KeyType. Added printRow() similar to printMap() to handle printing structs.

Reviewed By: DanielMunozT

Differential Revision: D38524066

fbshipit-source-id: 30bdbb5bba762c55cfeb58ad5cba37b1eb305422