Add docusaurus website skeleton (#2281)
Summary:
Adding the skeleton for a website based on docusaurus. For now the website is mostly based on the default template with a few changes to the color scheme and Velox icons, and pointers to github and current documentation websites.

It also adds a blog with an example monthly update post imported.

This is just the initial version so we can get the build system in place and configure Netlify; more content will be added in a follow up PRs.

Screenshot image of the website below:

![Screenshot from 2022-08-11 22-55-55](https://user-images.githubusercontent.com/12956990/184293394-ff5fa698-5918-41ca-8e78-070d28211b41.png)

and dark version:

![Screenshot from 2022-08-11 22-57-09](https://user-images.githubusercontent.com/12956990/184293462-db64e560-881b-465c-8b57-e912b20e09a3.png)

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2281

Reviewed By: kgpai

Differential Revision: D38670865

Pulled By: pedroerp

fbshipit-source-id: 12e0dcab2506e2b16111ab04f5d2354e928a95b2