Cleanup use of aggregates_ in HashTable. (#2339)
Summary:
The Hashtable maintains a list of aggregate functions for 2 purposes :

i) These functions were pass-through to the RowContainer which needed
them for sizing rows.

ii) The HashTable initialized new groups in these aggregates when it
detected new keys during lookups. These lookups only happened through the GroupingSet.

The aggregate initialization for new groups is better in the GroupingSet
logic than HashTable. The HashTable is then limited to do pure lookups and
sets up the required info in the HashLookup for use by the Operator.

HashTable continues to pass through the aggregates to the RowContainer only. So
we can remove this member variable also.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2339

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D38908583

Pulled By: mbasmanova

fbshipit-source-id: 826cdac2c6063102636f643fd8b74ef9e182bb94