Add simple in-memory Window operator. (#2157)
Summary:
This PR adds a simple in-memory Window operator. The row_number window function is implemented to test many usages of the Window operator.

Caveats:
i) This code has a very naive implementation of the Window operator. All input rows are sorted with a combination of (partitionkeys + orderbykeys) to bring the rows of each partition adjacent to each other and sorted by the order needed for the window function computation. This can be done in more performant ways, using a HashTable or otherwise. This work is pending.
ii) This code only supports the default window frame.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2157

Reviewed By: Yuhta

Differential Revision: D38954815

Pulled By: mbasmanova

fbshipit-source-id: 332da368890d2ef84a8478ba97a2e453f93e60c5