Propagate Errors from AsyncSource (#2421)
Summary:
Adds error propagation from AsyncSource. If the function of the source
produces an error, the source is considered to have a value and
accessing the value with move() will throw the same error on every
call. move() still returns null if another thread is already waiting
for the source. The waiting thread is still sure to get the error.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2421

Reviewed By: kgpai

Differential Revision: D39154416

Pulled By: oerling

fbshipit-source-id: 41bf716520a0f7978ad9f69d0db049b26cf49fcc