Fix loss of precision in timestamp when spilling (#2279)
Summary:
Currently we use a presto compatible serializer when spilling which
serializes timestamp to millisecond precision. This results in loss
of precision for velox timestamp type since velox stores it in
nanosecond precision. This patch ensures that precision is maintained
when spilling occurs. It introduces a new 'Options' input parameter
to the VectorSerde interface which is a map of key-value pairs. This
allows the PrestoSerializer to support serialization with both
precisions by allowing the caller to toggle between the two via the
input parameter.

Additionally, this also removes a hack inside the test suite which
by default downgrades timestamp precision to millisecond when
comparing results in order to be compatible with duckDB which is
used to verify correctness in some tests.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2279

Test Plan: Added unit tests

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D38644631

Pulled By: bikramSingh91

fbshipit-source-id: c53e36e8c05d3379722d83ef66620c1422aff70d