SimpleMemoryTracker that basically wraps an atomic
Summary:
When using memory usage tracker in trabant, we discovered that over time there are memory "leaks" due to different assumptions about the allocation patterns of big buffers.

While we track the accounting issue from the actual memory tracker, we acknowledge that a simpler implementation is a good short term solution to unblock the trabant integration.

Reviewed By: nazanincal

Differential Revision: D37293081

fbshipit-source-id: 4e54d6e9a0e798b9aaf65386d0caccdfbb355b3a