Fix bison and flex dependencies in MacOS CI (#2442)
Summary:
Bison and flex come standard with MacOS.  Out of caution, homebrew will not install the executables from these kegs into the bin directory on MacOS.

Due to this, the versions of bison and flex we've been using in CircleCI on MacOS have been the ones installed with the OS not the ones we installed with homebrew.

This change fixes the PATH we use when building to include the directories where homebrew installs the executables for these two libraries.  It also updates the FLEX_INCLUDE_DIR environment variable used in cmake to point to the include directory for flex installed by homebrew.

I think we've been getting away with this because the versions that came with our MacOS image were new enough that they worked for our build, but now our dependencies require newer versions of the libraries.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2442

Test Plan: Verified this fixes the MacOS CI build

Reviewed By: kgpai

Differential Revision: D39231396

Pulled By: kevinwilfong

fbshipit-source-id: d26ee523f52e45624d57aa69ccd5d73f18beccec