pass context and result by reference to apply() (#2376)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2376

instead of passing a pointer to VectorPtr and a pointer to the context
the change passes those by ref since they should never be passed as null.

For the same reason, the diff also makes the result vector argument of ensureWritable a reference instead of a pointer to pointer and also
constructor arguments of LocalDecodedVectors.

Reviewed By: kagamiori

Differential Revision: D38968157

fbshipit-source-id: aabd0fc2bee1694838d529fe6fda2624f234fda0