Fix ExchangeClient not properly release ExchangeQueue issue upon close (#2443)
Summary:
When Task terminates and cleans up resource, close() is called for ExchangeClient. But close() omits to gracefully clean up resources in ExchangeQueue when there is a MemoryPool provided in SerializedPage.

Crash will happen when there is still unprocessed SerializedPage in ExchangeQueue in ExchangeClient when Task is destructed. This is because of the destruction order of fields in Task: MemoryPool is destructed before ExchangeClient, making ExchangeClient not able to properly use MemoryPool to cleanup upon the invocation of its destructor.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2443

Reviewed By: mbasmanova

Differential Revision: D39228144

Pulled By: tanjialiang

fbshipit-source-id: e70ba4cc5cc288a10ba3526d91d77d96a7539b33