Add isVectorWritable function to BaseVector (#2232)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2232

This diff adds isVectorWritable to BaseVector.  This is a utility function that checks
that it is safe to write to the Vector.

It checks the following conditions
* The vector is singly referenced.
* The vector has a Flat-like encoding (Flat, Array, Map, Row).
* Any child Buffers are mutable (this also means they are singly referenced).
* All of these conditions hold for child Vectors recursively.

As part of this, I also modified isNullsWritable.  Previously, it would simply return
true if the Vector was not a Constant or Sequence Vector.  To be consistent both
with the naming and the way it is used, it now also checks that the nulls Buffer is
mutable.

Removed isReusableFlatVector and isRecyclable as these are just isVectorWritable with slightly stricter requirements that are easily specified inline where needed.

I plan to leverage this new function in
D38291699
https://github.com/facebookincubator/velox/pull/2167
in an assert to verify ensureWritable does not need to be called.

Reviewed By: mbasmanova

Differential Revision: D38515249

fbshipit-source-id: 3035247121efe6289d74e03be98c83839cd8c80c