Fix rare 'Output buffers for task not found' error (#2488)
Summary:
Occasionally, Prestissimo query using broadcast shuffle fails with an error:

VeloxRuntimeError: it != buffers.end() Output buffers for task not found: 20220907_072939_02883_iuzrj.3.0.4
     at Unknown.# 2  facebook::velox::exec::PartitionedOutputBufferManager::getBuffer
     at Unknown.# 3  facebook::velox::exec::PartitionedOutputBufferManager::updateBroadcastOutputBuffers
     at Unknown.# 4  facebook::velox::exec::Task::updateBroadcastOutputBuffers
     at Unknown.# 5  facebook::presto::TaskManager::createOrUpdateTask

I'm not able to reproduce this, but I believe this happens when coordinator
sends a duplicate updateBroadcastOutputBuffers with noMoreBuffers set to true.
The first call makes the task complete and the buffers being deleted. The
second call hits "Output buffers for task not found".

The fix is to ignore updateBroadcastOutputBuffers messages after receiving
no-more-buffers message.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/2488

Reviewed By: spershin

Differential Revision: D39372867

Pulled By: mbasmanova

fbshipit-source-id: d62b2f35e3b58da6219f4219979370255fcdaa5f