Fix TSAN errors in MmapAllocator (#2482)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/2482

TSAN was showing two sources of issues in MappedMemoryTest

The statistics variables in MmapAllocator are not atomic and written outside of a
mutex.  Based on the comment this looks like it was intentional, although the
reasoning wasn't documented.  I assume the reasoning was for performance
since their correctness won't impact the correctness of the program.  It looks like
everywhere they're updated other atomics are read/updated repeatedly, so I don't
think it should significantly hurt performance.  I don't see anywhere these
statistics are ever used so we can also safely remove them if we don't plan to use
them.

In the test itself we were allocating elements in a vector as we were spinning up
threads that read from the vector.  Although we were accessing the data in a
thread safe way, the size of the vector was being read/written in different threads
which was detected as a data race.  Creating the vector before spinning up the
threads fixes it.

Reviewed By: xiaoxmeng

Differential Revision: D39357067

fbshipit-source-id: 0b2ab037c59004e1bff8a98c085f556f6de65c2b